Enable Event Source tracing in tests optionally
