+++ src/Microsoft.Data.SqlClient/tests/ManualTests/DataCommon/DataTestUtility.cs
@@ -3,16 +3,13 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlTypes;
+using System.Diagnostics.Tracing;
 using System.Globalization;
 using System.IO;
-using System.Linq;
-using System.Reflection;
 using System.Security;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Identity.Client;
@@ -38,6 +35,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
         public static List<string> AEConnStrings = new List<string>();
         public static List<string> AEConnStringsSetup = new List<string>();
         public static readonly bool EnclaveEnabled = false;
+        public static readonly bool TracingEnabled = false;
         public static readonly bool SupportsIntegratedSecurity = false;
         public static readonly bool SupportsLocalDb = false;
         public static readonly bool SupportsFileStream = false;
@@ -56,6 +54,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                                                                      ".database.chinacloudapi.cn"};
 
         private static Dictionary<string, bool> AvailableDatabases;
+        private static TraceEventListener TraceListener;
 
         private class Config
         {
@@ -70,6 +69,7 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
             public string AzureKeyVaultClientId = null;
             public string AzureKeyVaultClientSecret = null;
             public bool EnclaveEnabled = false;
+            public bool TracingEnabled = false;
             public bool SupportsIntegratedSecurity = false;
             public bool SupportsLocalDb = false;
             public bool SupportsFileStream = false;
@@ -99,6 +99,12 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                 SupportsIntegratedSecurity = c.SupportsIntegratedSecurity;
                 SupportsFileStream = c.SupportsFileStream;
                 EnclaveEnabled = c.EnclaveEnabled;
+                TracingEnabled = c.TracingEnabled;
+
+                if(TracingEnabled)
+                {
+                    TraceListener = new DataTestUtility.TraceEventListener();
+                }
 
                 if (IsAADPasswordConnStrSetup() && IsAADAuthorityURLSetup())
                 {
@@ -636,6 +642,25 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                 }
             }
             return res;
+        }
+
+        public class TraceEventListener : EventListener
+        {
+            public List<int> IDs = new List<int>();
+
+            protected override void OnEventSourceCreated(EventSource eventSource)
+            {
+                if (eventSource.Name.Equals("Microsoft.Data.SqlClient.EventSource"))
+                {
+                    // Collect all traces for better code coverage
+                    EnableEvents(eventSource, EventLevel.Informational, EventKeywords.All);
+                }
+            }
+
+            protected override void OnEventWritten(EventWrittenEventArgs eventData)
+            {
+                IDs.Add(eventData.EventId);
+            }
         }
     }
 }
