+++ src/Microsoft.Data.SqlClient/tests/ManualTests/config.default.json
@@ -5,6 +5,7 @@
     "TCPConnectionStringAASVBS": "",
     "TCPConnectionStringAASSGX": "",
     "EnclaveEnabled": false,
+    "TracingEnabled": false,
     "AADAuthorityURL": "",
     "AADPasswordConnectionString": "",
     "AzureKeyVaultURL": "",
